workshops/may-2025/workshop3/semantic-search/requirements.txt
openai>=1.0.0
chromadb>=0.4.18
python-dotenv>=1.0.0